+++ braintree/test/nonces.py
@@ -60,6 +60,7 @@ class Nonces(object):
     PayPalFuturePaymentRefreshToken = "fake-paypal-future-refresh-token-nonce"
     SEPA = "fake-sepa-bank-account-nonce"
     GatewayRejectedFraud = "fake-gateway-rejected-fraud-nonce"
+    GatewayRejectedRiskThreshold = "fake-gateway-rejected-risk-thresholds-nonce"
     MasterpassAmEx = "fake-masterpass-amex-nonce"
     MasterpassDiscover = "fake-masterpass-discover-nonce"
     MasterpassMasterCard = "fake-masterpass-mastercard-nonce"